+++ sketch_oxide/src/membership/mod.rs
@@ -18,6 +18,7 @@ mod stable_bloom;
 mod stacked_filter;
 mod vacuum_filter;
 mod vector_quotient_filter;
+mod xor_filter;
 
 pub use adaptive_quotient_filter::AdaptiveQuotientFilter;
 pub use binary_fuse::BinaryFuseFilter;
@@ -35,6 +36,7 @@ pub use stable_bloom::StableBloomFilter;
 pub use stacked_filter::StackedFilter;
 pub use vacuum_filter::{VacuumFilter, VacuumFilterStats};
 pub use vector_quotient_filter::VectorQuotientFilter;
+pub use xor_filter::XorFilter;
 
 #[cfg(test)]
 mod tests {